requirements.txt
streamlit
tensorflow==2.7.0
numpy
matplotlib
Pillow